When Capsicum is enabled, use cap_enter() and POSIX Shared Memory.
We can't use do a full System V Shared Memory pipeline between the
renderer and X11 because the global sysvshm namespace is not
available from within the sandbox.  Instead, borrow the Mac OS X
shared memory code and use that; for now, we also disable shared
memory for pixmaps between the browser and the X server, but may
want to re-enable that later.